backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-restframework==0.0.1
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
